cell_1

This dataset is an example of a single-cell file that can be input to vmrseq::data.pool function. It contains the first 10000 CpG sites in a random cell from mouse frontal cortext dataset published by Luo et al. (2017).

data(cell_1)

cell_2

cell_3

Format

A data frame with 10000 rows and 5 variables (no column names):

V1

Chromosome

V2

Genomic coordinate

V3

Strand information

V4

Number of methylated reads

V5

Number of reads in total

An object of class data.table (inherits from data.frame) with 10000 rows and 5 columns.

An object of class data.table (inherits from data.frame) with 10000 rows and 5 columns.

References

Luo, Chongyuan et al. Single-cell methylomes identify neuronal subtypes and regulatory elements in mammalian cortex.. Science (New York, N.Y.) vol. 357,6351 (2017): 600-604.

Examples

data(cell_1)
cell_1
#>           chr      pos strand mc_count total
#>        <char>    <int> <char>    <int> <int>
#>     1:   chr1  3003885      *        1     1
#>     2:   chr1  3003898      *        1     1
#>     3:   chr1  3011550      *        1     1
#>     4:   chr1  3019920      *        1     1
#>     5:   chr1  3020794      *        2     2
#>    ---                                      
#>  9996:   chr1 38247225      *        1     1
#>  9997:   chr1 38248100      *        0     1
#>  9998:   chr1 38249879      *        1     1
#>  9999:   chr1 38249890      *        1     1
#> 10000:   chr1 38249944      *        1     1