Pool single-cell file together into an HDF5-based SummarizedExperiment object with sparse matrix representation.

This function pools individual-cell CpG read files into a SummarizedExperiment object so that it can be input to the vmrseq.smooth function. Note that in each cell, sites with hemimethylation or intermediate methylation levels (i.e., 0 < meth_read/total_read < 1) will be removed.

data.pool(
  cellFiles,
  sep,
  writeDir,
  chrNames,
  colData = NULL,
  sparseNAdrop = TRUE
)

Arguments

cellFiles

Vector of character strings indicating single-cell file paths you wish to pool into SE object(s). Cell files should be in BED-like format, where the first 5 columns in each file must be: <chr>, <pos>, <strand>, <meth_read>, <total_read>, in strict order. Each row contains info of one CpG. Strand is assumed to be properly filpped. Cell files can be zipped ones in .gz format.

sep

the field separator character. Values on each line of cell file are separated by this character.

writeDir

A single character string indicating a folder directory where you wish to store the processed SE object(s). The SE will be stored in HDF5 format.

chrNames

Single or a vector of character strings representing chromosome names in cell files. Only chromosomes listed in selectChrs will be processed.

colData

A DataFrame or data.frame object containing colData for the SE object (applied to all chromosomes).

sparseNAdrop

Logical value indicating whether or not to use NA-dropped sparseMatrix representation. 'NA-dropped' means replacing NAs as zeros and then represent 0 as a very small positive value close to 0 so that the data matrix is converted to a traditional sparseMatrix form with 0 as default entry value. (see sparseNAMatrix-class from package recommenderlab). Using NA-dropped sparseMatrix representation helps to save RAM during data processing, as well as storage space saving to disk. Currently only supports TRUE!!

Value

Directly write out to the `writeDir` and does not return anything.

Examples

if (FALSE) data.pool(cellFiles = your_cell_file_list, sep = ",", chrNames = "chr1", writeDir = "your/write/path") # \dontrun{}