Compute regional methylation information for individual cells.

This function summarize the methylated CpG count and total CpG count per region per cell.

region.summary(
  SE,
  region_ranges,
  sparseNAdrop = is_sparse(assays(SE)[[1]]),
  verbose = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

Arguments

SE

SummarizedExperiment object with one (and only one) assay that contains *binary* methylation status of CpG sites in individual cells. In usual analysis workflow (of vmrseq), SE should be the output of vmrseq::data.pool.

region_ranges

GRanges object that contains genomic coordinates of regions of interest.

sparseNAdrop

logical value that represents whether the NA values are droppped in the input SE object. SE objects output by vmrseq::data.pool are NA dropped. See ?vmrseq::data.pool for details about NA-dropped representation.

verbose

logical value that indicates whether progress messages should be printed to stdout. Defaults value is TRUE.

BPPARAM

a BiocParallelParam object to specify the parallel backend. The default option is BiocParallel::bpparam() which will automatically creates a cluster appropriate for the operating system.

Value

Returns a SummarizedExperiment object of dimension (# regions x # cells). In total the object contains thress assays. Specifically, `M` and `Cov` represent the number of methylated CpGs and the number of covered CpGs in each region per cell; and `MF` (stands for methylation fraction) represents the regional average methylation level computed by `M/Cov`.

Examples

# load example data
toy.se <- HDF5Array::loadHDF5SummarizedExperiment(system.file("extdata", "toy", package = "vmrseq"))
data(toy.results)

# run vmrseq.fit
regions.se <- region.summary(SE = toy.se, region_ranges = toy.results$vmr.ranges[1:3])
#> Parallel: Parallelizing using 2 workers/cores (backend: BiocParallel:MulticoreParam).
regions.se
#> class: RangedSummarizedExperiment 
#> dim: 3 200 
#> metadata(0):
#> assays(3): M Cov MF
#> rownames: NULL
#> rowData names(5): num_cpg start_ind end_ind pi loglik_diff
#> colnames: NULL
#> colData names(0):