Construct candidate regions and detect variably methylated regions.

Construct candidate regions (CRs) by taking groups of consecutive loci that exceed threshold on the variance of smoothed relative methylation levels and detect variably methylated regions (VMRs) by optimizing a hidden Markov model (HMM).

vmrseq.fit(
  gr,
  alpha = 0.05,
  maxGap = 2000,
  stage1only = FALSE,
  minNumCR = 5,
  minNumVMR = 5,
  gradient = TRUE,
  tp = NULL,
  control = vmrseq.optim.control(),
  verbose = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

Arguments

gr

GRanges object output by vmrseq::vmrseq.smooth, containing genomic coordinates (chr, start, end) and summarized information (meth, total, var) of CpG sites in the input dataset.

alpha

positive scalar value between 0 and 1 that represents the designated significance level for determining variance threshold of candidate regions construction. The variance threshold is determined by taking 1-alpha quantile value of an approximate null distribution of variance simulated from the beta priors of emission probability in the hidden Markov model. Default value of alpha is 0.05.

maxGap

integer value representing maximum number of base pairs in between neighboring CpGs to be included in the same VMR. Default value is 2000 bp.

stage1only

boolean value indicating whether the algorithm should run stage 1 of vmrseq (the construction of candidate regions) only. If set to TRUE, the function will output only the candidate regions. Default is FALSE.

minNumCR

positive integer value representing the minimum number of CpG sites within a candidate region. Default value is 5.

minNumVMR

positive integer value representing the minimum number of CpG sites within a variably methylated region. Default value is 5.

gradient

logical value indicating whether exponentiated gradient descent shall be applied to update prevalence parameter. Default is TRUE. If set as FALSE, initial values (i.e., value of inits arguments in vmrsqe::vmrseq.optim.control, can be set up in the control argument of this function) are used as prevalence parameter for decoding hidden states.

tp

a `transitProbs-class` object that contains the transition probability distribution used for HMM optimization. Default value is transition probability vmrseq:::tp0 built in the package that was previously trained on mouse brain cells. See manuscript for training procedure and data source.

control

list of miscellaneous parameters used to control optimization of the HMM model. Default is output of vmrseq::vmrseq.optim.control(). Can be changed by tweaking arguments in function vmrseq::vmrseq.optim.control().

verbose

logical value that indicates whether progress messages should be printed to stdout. Defaults value is TRUE.

BPPARAM

a BiocParallelParam object to specify the parallel backend. The default option is BiocParallel::bpparam() which will automatically creates a cluster appropriate for the operating system.

Value

The results object is a list of 6 elements that contains the following information: 1. `gr`: The `Granges` object that has been input to `vmrseq.fit` with two added metadata columns: + `cr_index` = Index in reference to rows of `cr.ranges`, denoting row number of the candidate region to which the CpG site belongs. + `vmr_index` = Index in reference to rows of `vmr.ranges`, denoting row number of the variably methylated region to which the CpG site belongs. 2. `vmr.ranges`: A `Granges` object with the coordinates of each detected variably methylated region (each row is a VMR), with metadata columns: + `num_cpg` = Number of observed CpG sites in the VMR. + `start_ind` = Index of the starting CpG sites in reference to rows of `gr`. + `end_ind` = Index of the ending CpG sites in reference to rows of `gr`. + `pi` = Prevalence of the methylated grouping (see manuscript for details) + `loglik_diff` = Difference in log-likelihood of two-grouping and one-grouping HMM fitted to the VMR; can be used to rank the VMRs. 3. `cr.ranges`: A `Granges` object with the coordinates of each candidate region (each row is a candidate region), with metadata column: + `num_cpg` = Number of observed CpG sites in the candidate region. 4. `alpha`: Designated significance level (default 0.05, can be changed by user with function argument). It is used for determining the threshold on variance used for constructing candidate. The threshold is computed by taking the (1-alpha) quantile of an approximate null distribution of variance (see manuscript for details). 5. `var_cutoff`: Variance cutoff computed from `alpha`. 6. `bb_params`: Beta-binomial parameter used in emission probability of the HMM model; they are determined by the magnitude of the input dataset (see manuscript for details).

Examples

# load example data
data(toy.gr)
# run vmrseq.fit
toy.results <- vmrseq.fit(toy.gr)
#> Parallel: Parallelizing using 2 workers/cores (backend: BiocParallel:MulticoreParam).
#> Step 1: Detecting candidate regions...
#> ...Calling candidate regions with cutoff of 0.117 on variance.
#> ...Finished calling candidate regions - found 73 candidate regions in total.
#>   ...3.25% sites are called to be in candidate regions.
#> Step 2: Detecting VMRs...
#> ...Finished detecting VMRs - took 0.2 min and 63 VMRs found in total.
#>   ...2.87% sites are called to be in VMRs.
toy.results
#> $gr
#> GRanges object with 29313 ranges and 5 metadata columns:
#>           seqnames    ranges strand |      meth     total        var  cr_index
#>              <Rle> <IRanges>  <Rle> | <integer> <numeric>  <numeric> <integer>
#>       [1]   pseudo   3000827      * |        17        17 0.00479461      <NA>
#>       [2]   pseudo   3001007      * |        19        19 0.00479461      <NA>
#>       [3]   pseudo   3001018      * |        16        16 0.00479461      <NA>
#>       [4]   pseudo   3001277      * |        14        15 0.00479461      <NA>
#>       [5]   pseudo   3001629      * |        16        16 0.00479461      <NA>
#>       ...      ...       ...    ... .       ...       ...        ...       ...
#>   [29309]   pseudo   8523533      * |        15        15  0.0149416      <NA>
#>   [29310]   pseudo   8523719      * |        11        11  0.0158558      <NA>
#>   [29311]   pseudo   8523783      * |         6         6  0.0158558      <NA>
#>   [29312]   pseudo   8523864      * |         7         7  0.0158558      <NA>
#>   [29313]   pseudo   8524231      * |         2         4  0.0199032      <NA>
#>           vmr_index
#>           <integer>
#>       [1]      <NA>
#>       [2]      <NA>
#>       [3]      <NA>
#>       [4]      <NA>
#>       [5]      <NA>
#>       ...       ...
#>   [29309]      <NA>
#>   [29310]      <NA>
#>   [29311]      <NA>
#>   [29312]      <NA>
#>   [29313]      <NA>
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $vmr.ranges
#> GRanges object with 63 ranges and 5 metadata columns:
#>        seqnames          ranges strand |   num_cpg start_ind   end_ind
#>           <Rle>       <IRanges>  <Rle> | <integer> <numeric> <numeric>
#>    [1]   pseudo 3121060-3122009      * |        13       674       686
#>    [2]   pseudo 3196918-3198634      * |        11      1066      1076
#>    [3]   pseudo 3217393-3221314      * |        21      1145      1165
#>    [4]   pseudo 3234201-3236719      * |        15      1210      1224
#>    [5]   pseudo 3252227-3259375      * |        33      1285      1317
#>    ...      ...             ...    ... .       ...       ...       ...
#>   [59]   pseudo 7429380-7430593      * |         7     24062     24068
#>   [60]   pseudo 7505596-7507387      * |         7     24365     24371
#>   [61]   pseudo 7671976-7672660      * |         9     25171     25179
#>   [62]   pseudo 8399348-8400680      * |        10     28758     28767
#>   [63]   pseudo 8410430-8411660      * |        32     28803     28834
#>               pi loglik_diff
#>        <numeric>   <numeric>
#>    [1]  0.526617    24.37579
#>    [2]  0.398940    15.66478
#>    [3]  0.817459     7.27884
#>    [4]  0.772000    10.02922
#>    [5]  0.423506    42.29019
#>    ...       ...         ...
#>   [59]  0.451126     8.20103
#>   [60]  0.362805     8.75095
#>   [61]  0.698666     9.35440
#>   [62]  0.224853     7.65353
#>   [63]  0.426745    41.88692
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $cr.ranges
#> GRanges object with 73 ranges and 1 metadata column:
#>        seqnames          ranges strand |   num_cpg
#>           <Rle>       <IRanges>  <Rle> | <integer>
#>    [1]   pseudo 3121060-3122009      * |        13
#>    [2]   pseudo 3196918-3198634      * |        11
#>    [3]   pseudo 3217393-3221314      * |        21
#>    [4]   pseudo 3234201-3236719      * |        15
#>    [5]   pseudo 3252227-3259375      * |        33
#>    ...      ...             ...    ... .       ...
#>   [69]   pseudo 7671976-7672711      * |        12
#>   [70]   pseudo 7800327-7801784      * |         5
#>   [71]   pseudo 8399348-8400680      * |        10
#>   [72]   pseudo 8410068-8411660      * |        33
#>   [73]   pseudo 8495540-8497541      * |         5
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $alpha
#> [1] 0.05
#> 
#> $var_cutoff
#>       95% 
#> 0.1169572 
#> 
#> $bb_params
#> $bb_params$pars_u
#>        nu        mu     sigma 
#> 0.4131979 0.1871845 0.1205215 
#> 
#> $bb_params$pars_m
#>         mu      sigma 
#> 0.91401348 0.09909785 
#> 
#>