Smoothing on single-cell bisulfite sequencing data for the purpose of constructing candidate regions.

vmrseq.smooth takes a SummarizedExperiment object with information of methylation level of individual cells as input, and perform a kernel smoother to ‘relative’ methylation levels of individual cells prior to constructing candidate regions. Purpose of the smoothing is to adjust for uneven coverage biases and borrow information from nearby sites. See manuscript for detailed description.

vmrseq.smooth(
  SE,
  bpWindow = 2000,
  sparseNAdrop = is_sparse(assays(SE)[[1]]),
  verbose = TRUE,
  BPPARAM = bpparam()
)

Arguments

SE

SummarizedExperiment object with one (and only one) assay that contains *binary* methylation status of CpG sites in individual cells. We recommend using output by vmrseq::data.pool (i.e., an NA-dropped HDF5-based SummarizedExperiment object) to prevent running out of memory.

bpWindow

positive integer that represents the width (in bp) of smoothing window. Default value is 2000.

sparseNAdrop

logical value that represents whether the NA values are droppped in the input SE object. SE objects output by vmrseq::data.pool are NA dropped. See ?vmrseq::data.pool for details about NA-dropped representation.

verbose

logical value that indicates whether progress messages should be printed to stdout. Defaults value is TRUE.

BPPARAM

a BiocParallelParam object to specify the parallel backend. The default option is BiocParallel::bpparam() which will automatically creates a cluster appropriate for the operating system.

Value

a GRanges object that contains the result of smoothing. The object retains genomic coordinates (chr, start, end) of input CpG sites, in the same order as in the input SE object. Three column are added (on top of original metadata columns for the CpG sites in SE, if any): 1. meth: methylated cell count of the CpG 2. total: total (non-missing) cell count of the CpG 3. var: variance computed based on individual-cell smoothed relative methylation levels.

See also

Examples

# load example data
toy.se <- HDF5Array::loadHDF5SummarizedExperiment(system.file("extdata", "toy", package = "vmrseq"))

# preprocessing
total <- DelayedArray::rowSums(SummarizedExperiment::assays(toy.se)$M_mat > 0)
toy.se <- subset(toy.se, total >= 3)

# run vmrseq.smooth
toy.gr <- vmrseq.smooth(toy.se)
#> Parallel: Parallelizing using 2 workers/cores (backend: BiocParallel:MulticoreParam).
#> Smoothing in progress...
#> ...Chromosome pseudo: 
#> Variance computed (0.06 min). 
toy.gr
#> GRanges object with 29313 ranges and 3 metadata columns:
#>           seqnames    ranges strand |      meth     total        var
#>              <Rle> <IRanges>  <Rle> | <integer> <numeric>  <numeric>
#>       [1]   pseudo   3000827      * |        17        17 0.00479461
#>       [2]   pseudo   3001007      * |        19        19 0.00479461
#>       [3]   pseudo   3001018      * |        16        16 0.00479461
#>       [4]   pseudo   3001277      * |        14        15 0.00479461
#>       [5]   pseudo   3001629      * |        16        16 0.00479461
#>       ...      ...       ...    ... .       ...       ...        ...
#>   [29309]   pseudo   8523533      * |        15        15  0.0149416
#>   [29310]   pseudo   8523719      * |        11        11  0.0158558
#>   [29311]   pseudo   8523783      * |         6         6  0.0158558
#>   [29312]   pseudo   8523864      * |         7         7  0.0158558
#>   [29313]   pseudo   8524231      * |         2         4  0.0199032
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths